requirements.txt
streamlit
opencv-python
numpy
keras
Pillow
tensorflow